fix: datetimes for meetingtimes
